Make listen URIs configurable
Previously, slapd in this container would always listen on only
`ldaps:///` and `ldap:///`. As indicated in HACKING.md, this makes it
difficult to make configuration changes to `cn=config`.  With this
change, the user may set OPENLDAP_LISTEN_URIS to change the list of
endspoints to which slapd will listen, thus permitting the use of
`ldapi:///` if so desired.  E.g. `docker run -e
OPENLDAP_LISTEN_URIS="ldapi/// ldaps:///" ...`.

This also restricts slapd to `ldapi:///` when performing the initial
configuration steps.